+++ lib/src/model/storage_pre_model.dart
@@ -1,0 +1,15 @@
+class View360ChatPrefsModel {
+  final String chatId;
+  final String customerId;
+  final String customerName;
+  final String customerEmail;
+  final String customerPhone;
+
+  View360ChatPrefsModel({
+    required this.chatId,
+    required this.customerId,
+    required this.customerName,
+    required this.customerEmail,
+    required this.customerPhone,
+  });
+}